// ClientApp/src/app/views/reporting/models/employee-list-reporting.ts
export class EmployeeListReporting {
    departmentId: any ;
    departmentName: any ;
    sectionId: any ;
    sectionName: any ;
    total: any ;
    allTotal: any ;
    empId: any ;
    idCardNo: any ;
    empName: any ;
    designationId: any ;
    designationName: any ;
    mobile: any ;
    joinDate: any ;
}